Created logout button and created sales payload models (Have to fix API call)

+++ lib/screens/create_sale/create_sale_screen.dart
@@ -7,9 +7,6 @@ import 'package:play_hq/helpers/app_enums.dart';
 import 'package:play_hq/helpers/app_fonts.dart';
 import 'package:play_hq/helpers/app_screen_utils.dart';
 import 'package:play_hq/screens/create_sale/widgets/create_sale_confirm_bottom_sheet_widget.dart';
-import 'package:play_hq/screens/create_sale/widgets/sale_confirm_dialog.dart';
-import 'package:play_hq/service_locator.dart';
-import 'package:play_hq/services/dialog_service.dart';
 import 'package:play_hq/view_models/sales/create_sale/create_sale_model.dart';
 import 'package:play_hq/widgets/custom_body.dart';
 import 'package:play_hq/widgets/custom_button_widget.dart';
@@ -117,7 +114,7 @@ class _CreateSaleScreenState extends State<CreateSaleScreen> {
                             hintText: "Enter the price",
                             suffixText: 'LKR',
                             onChanged: (value) => value != ""
-                                ? model.setPrice(double.parse(value))
+                                ? model.setPrice(int.parse(value))
                                 : model.setPrice(0),
                           );
                         }),
